Remove instructions template column
